#5 Refactored structure of directories.

+++ city-catalog/src/main/java/com/easymoto/city/CityRepository.java
@@ -5,12 +5,14 @@ import java.util.List;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.Repository;
 
+import com.easymoto.city.domain.City;
+
 /**
  * Repository for City data implemented using Spring Data JPA.
  * 
  * @author Carlos E. Santin <cesantin@gmail.com>
  */
-public interface CityRepository extends Repository<City, Long> {
+public interface CityRepository extends Repository<City, Integer> {
   /**
    * Find an city with the specified id.
    *
